+++ src/activities/ActivityManager.h
@@ -11,6 +11,7 @@
 #include <vector>
 
 #include "GfxRenderer.h"
+#include "CrossPointSettings.h"
 #include "MappedInputManager.h"
 #include "util/ScreenshotInfo.h"
 
@@ -134,6 +135,7 @@ class ActivityManager {
 #endif
   bool canSnapshotForSleepOverlay() const;
   bool requestManualReaderRefresh();
+  bool handleShortcutAction(CrossPointSettings::SHORT_PWRBTN action);
   bool skipLoopDelay() const;
   std::string getCurrentBookPath() const;
   ScreenshotInfo getScreenshotInfo() const;